% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{filter_data}
\alias{filter_data}
\title{Filter Genome Comparator data}
\usage{
filter_data(data, gene_freq_cutoff = 0.2, missing_gene_cutoff = 0.8, out = NA)
}
\arguments{
\item{data}{A data frame of Genome Comparator output with disease as a column}

\item{gene_freq_cutoff}{Genes in less than this proportion of isolates in the dataset will be removed}

\item{missing_gene_cutoff}{Genes missing in more than this many isolates will be removed}

\item{out}{Filepath to a directory for writing output files}
}
\description{
Removes rubbish data...
}
