% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_imp_feats.R
\name{get_imp_feats}
\alias{get_imp_feats}
\title{Identify and plot important predictors}
\usage{
get_imp_feats(model, annotate = F, annotations, topimportant = 20, out)
}
\arguments{
\item{model}{A model object}

\item{annotate}{Logical. Annotate PubMLST ids or plot the ids themselves}

\item{annotations}{A dataframe consisting of PuBMLST ids and the annotations to convert them to}

\item{topimportant}{Number of features to plot}

\item{out}{File path to write output files to}
}
\description{
Identify and plot important predictors
}
