% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_synonymous.R
\name{merge_synonymous}
\alias{merge_synonymous}
\title{Merges ids of synonymous alleles}
\usage{
merge_synonymous(df, allele_dir, out)
}
\arguments{
\item{df}{A dataframe object containing allele ids for a set of genes akin to Genome Comparator output from PuBMLST. Should already be filtered with filter_data()}

\item{allele_dir}{A directory containing a set of fasta files, each of which contain all the allele sequences for a given gene. Allele ids in df will be matched to sequences in these files.}

\item{out}{The path and filename to write a .xlsx file of the output to.}
}
\description{
Merges ids of synonymous alleles
}
