% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_model.R
\name{opt_model}
\alias{opt_model}
\title{Optimise a random forest model}
\usage{
opt_model(data, params, ds_fac_train)
}
\arguments{
\item{data}{input training dataset}

\item{params}{params to test}

\item{ds_fac_train}{Downsampling factor for training data by which to modify class weights for model training}
}
\description{
Optimise a random forest model
}
