% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{Predict disease}
\usage{
run_model(train, test, params, importance = importance, out, ds_fac_train)
}
\arguments{
\item{train}{Training data}

\item{test}{Test data}

\item{params}{Optimised parameters for the model to use}

\item{importance}{Importance metric - either 'impurity' or 'permutation'}

\item{out}{filepath for output files to be written to}

\item{ds_fac_train}{downsampling factor for the training data, by which the class weights will be modified.}
}
\description{
Predict disease
}
