% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_feats.R
\name{select_feats_aov}
\alias{select_feats_aov}
\title{Reduce feature space with univariate anova selection
Calculates ANOVA-derived p values for each feature against the predictor and keeps only the most significant features}
\usage{
select_feats_aov(data, size, out)
}
\arguments{
\item{data}{\itemize{
\item data frame of filtered Genome Comparator output used as training data.
}}

\item{size}{Number of the most significant features to keep}

\item{out}{File path to a directory to save files to}
}
\description{
Reduce feature space with univariate anova selection
Calculates ANOVA-derived p values for each feature against the predictor and keeps only the most significant features
}
