% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_feats.R
\name{select_feats_Bor}
\alias{select_feats_Bor}
\title{Reduce feature space with Boruta}
\usage{
select_feats_Bor(df, maxRuns = 1000, out)
}
\arguments{
\item{df}{Input data frame of filtered Genome Comparator output.}

\item{maxRuns}{Maximum number of feature selection iterations for Boruta to perform}

\item{out}{File path to a directory to save files to}
}
\description{
Reduce feature space with Boruta
}
