% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_model.R
\name{set_params}
\alias{set_params}
\title{Define the parameter space for model optimisation}
\usage{
set_params(
  n_trees_min = 500,
  n_trees_max = 5000,
  n_trees_inc = 500,
  mtry_min = 5,
  mtry_max = 30,
  mtry_inc = 1,
  nodesize_min = 1,
  nodesize_max = 50,
  nodesize_inc = 5,
  searchsize = 100
)
}
\arguments{
\item{n_trees_min}{Min number of trees}

\item{n_trees_max}{Max number of trees}

\item{n_trees_inc}{Increment for trees to test}

\item{mtry_min}{Min number of mtry}

\item{mtry_max}{Max number of mtry}

\item{mtry_inc}{Increment for mtry values to test}

\item{nodesize_min}{Min node size}

\item{nodesize_max}{Max node size}

\item{nodesize_inc}{Increment for node size values to test}

\item{searchsize}{Number of searches in the grid to perform.}
}
\description{
Define the parameter space for model optimisation
}
