% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{split_data}
\alias{split_data}
\title{Split data into train and test datasets}
\usage{
split_data(df, train_proportion = 0.7, balance_test_data = FALSE, out)
}
\arguments{
\item{df}{Input data frame of filtered and feature selected Genome Comparator output}

\item{train_proportion}{Proportion of the dataset to use for training. The remainder will be used as test data}

\item{balance_test_data}{Should the proportion of carriage and disease isolates in the test dataset be equalised?}

\item{out}{Filepath to write files to}
}
\description{
Split data into train and test datasets
}
